Finished wiht observations. Observations n/Smaller schools and charter schools have better average scores and passing rates
In both these scenarios the average scores are not as far off as the passing rates meaning that there
is a wider range of scores in the bigger schools and distric schools

Schools with more money per student does not make them more successful based on average scores or
passing rates. Meaning there are more factors that lead to a schools success than money.